requirements.txt
PyYAML==3.11
argparse==1.2.1
boto==2.38.0
distribute==0.6.24
wsgiref==0.1.2
